ADD More fields to MealDetailView
